pyproject.toml
[tool.poetry]
name = "onbot"
version = "0.1.0"
description = ""
authors = ["Suhrob Malikov <oopanndaa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.11.2"
tortoise-orm = "^0.16.18"
python-dotenv = "^0.15.0"
asyncpg = "^0.21.0"
psycopg2-binary = "^2.8.6"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
